requirements.txt
fastapi
uvicorn[standard]
requests
fake-useragent
python-dateutil
python-dotenv